Removed cascades. Removed gateway interfaces. fixed tests partly
