+++ src/main/java/com/knighttodo/knighttodo/gateway/DayTodoGateway.java
@@ -1,20 +1,43 @@
 package com.knighttodo.knighttodo.gateway;
 
 import com.knighttodo.knighttodo.domain.DayTodoVO;
+import com.knighttodo.knighttodo.gateway.privatedb.mapper.DayTodoMapper;
+import com.knighttodo.knighttodo.gateway.privatedb.repository.DayTodoRepository;
+import com.knighttodo.knighttodo.gateway.privatedb.representation.DayTodo;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
-public interface DayTodoGateway {
+@RequiredArgsConstructor
+@Service
+public class DayTodoGateway {
 
-    DayTodoVO save(DayTodoVO dayTodoVO);
+    private final DayTodoRepository dayTodoRepository;
+    private final DayTodoMapper dayTodoMapper;
 
-    List<DayTodoVO> findAll();
+    public DayTodoVO save(DayTodoVO dayTodoVO) {
+        DayTodo savedDayTodo = dayTodoRepository.save(dayTodoMapper.toTodo(dayTodoVO));
+        return dayTodoMapper.toTodoVO(savedDayTodo);
+    }
 
-    Optional<DayTodoVO> findById(UUID todoId);
+    public List<DayTodoVO> findAll() {
+        return dayTodoRepository.findAll().stream().map(dayTodoMapper::toTodoVO).collect(Collectors.toList());
+    }
 
-    void deleteById(UUID todoId);
+    public Optional<DayTodoVO> findById(UUID todoId) {
+        return dayTodoRepository.findById(todoId).map(dayTodoMapper::toTodoVO);
+    }
 
-    List<DayTodoVO> findByDayId(UUID dayId);
+    public void deleteById(UUID todoId) {
+        dayTodoRepository.deleteById(todoId);
+    }
+
+    public List<DayTodoVO> findByDayId(UUID dayId) {
+        return dayTodoRepository.findByDayId(dayId).stream().map(dayTodoMapper::toTodoVO)
+                .collect(Collectors.toList());
+    }
 }
