+++ src/main/java/com/knighttodo/knighttodo/gateway/RoutineTodoInstanceGateway.java
@@ -1,20 +1,46 @@
 package com.knighttodo.knighttodo.gateway;
 
 import com.knighttodo.knighttodo.domain.RoutineTodoInstanceVO;
+import com.knighttodo.knighttodo.gateway.privatedb.mapper.RoutineTodoInstanceMapper;
+import com.knighttodo.knighttodo.gateway.privatedb.repository.RoutineTodoInstanceRepository;
+import com.knighttodo.knighttodo.gateway.privatedb.representation.RoutineTodoInstance;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
-public interface RoutineTodoInstanceGateway {
+@RequiredArgsConstructor
+@Service
+public class RoutineTodoInstanceGateway {
 
-    RoutineTodoInstanceVO save(RoutineTodoInstanceVO routineTodoInstanceVO);
+    private final RoutineTodoInstanceRepository routineTodoInstanceRepository;
+    private final RoutineTodoInstanceMapper routineTodoInstanceMapper;
 
-    List<RoutineTodoInstanceVO> findAll();
+    public RoutineTodoInstanceVO save(RoutineTodoInstanceVO routineTodoInstanceVO) {
+        RoutineTodoInstance savedRoutineTodoInstance = routineTodoInstanceRepository
+                .save(routineTodoInstanceMapper.toRoutineTodoInstance(routineTodoInstanceVO));
+        return routineTodoInstanceMapper.toRoutineTodoInstanceVO(savedRoutineTodoInstance);
+    }
 
-    Optional<RoutineTodoInstanceVO> findById(UUID routineTodoInstanceId);
+    public List<RoutineTodoInstanceVO> findAll() {
+        return routineTodoInstanceRepository.findAll().stream()
+                .map(routineTodoInstanceMapper::toRoutineTodoInstanceVO).collect(Collectors.toList());
+    }
 
-    void deleteById(UUID routineTodoInstanceId);
+    public Optional<RoutineTodoInstanceVO> findById(UUID routineTodoInstanceId) {
+        return routineTodoInstanceRepository.findById(routineTodoInstanceId).map(routineTodoInstanceMapper::toRoutineTodoInstanceVO);
+    }
 
-    List<RoutineTodoInstanceVO> findByRoutineId(UUID routineInstanceId);
+    public void deleteById(UUID routineTodoId) {
+        routineTodoInstanceRepository.deleteById(routineTodoId);
+    }
+
+    public List<RoutineTodoInstanceVO> findByRoutineId(UUID routineInstanceId) {
+        return routineTodoInstanceRepository.findByRoutineInstanceId(routineInstanceId).stream()
+                .map(routineTodoInstanceMapper::toRoutineTodoInstanceVO)
+                .collect(Collectors.toList());
+    }
 }
