+++ src/main/java/com/knighttodo/knighttodo/gateway/privatedb/representation/Day.java
@@ -26,6 +26,6 @@ public class Day {
     @Column(name = "day_name")
     private String dayName;
 
-    @OneToMany(mappedBy = "day", cascade = CascadeType.ALL)
+    @OneToMany(mappedBy = "day", cascade = CascadeType.MERGE)
     private List<DayTodo> dayTodos = new ArrayList<>();
 }
